+++ hydra_genetics/utils/io/hotspot_report.py
@@ -1,3 +1,4 @@
+import builtins
 import logging
 from collections import OrderedDict
 
@@ -63,7 +64,6 @@ def generate_hotspot_report(sample,
             other.append(variant)
     log.info("Open genomic vcf")
     g_variants = VariantFile(gvcf_file)
-    sample_format_index_mapper = {sample: index + 1 for index, sample in enumerate(g_variants.header.samples)}
 
     columns = {'columns': []}
     if column_yaml_file is not None:
@@ -72,7 +72,16 @@ def generate_hotspot_report(sample,
             columns = yaml.load(file, Loader=yaml.FullLoader)
 
     output_order = []
-    hotspot_columns = ['sample', 'chr', 'start', 'stop', 'ref', 'alt', 'report', 'gvcf_depth', 'ref_depth', 'alt_depth']
+    hotspot_columns = {'sample': {},
+                       'chr': {},
+                       'start': {},
+                       'stop': {},
+                       'ref': {},
+                       'alt': {},
+                       'report': {},
+                       'gvcf_depth': {},
+                       'ref_depth': {},
+                       'alt_depth': {}}
     report_header = []
 
     biggest_order = -1
@@ -81,9 +90,12 @@ def generate_hotspot_report(sample,
     if isinstance(columns['columns'], dict):
         gcvf_depth_field = columns['columns'].get('gvcf_depth', {}).get('field', 'DP')
     for entry in columns['columns']:
+        if entry in hotspot_columns:
+            if columns['columns'][entry].get('visible', 1) == 0 or 'order' in columns['columns'][entry]:
+                del hotspot_columns[entry]
+            elif columns['columns'][entry].get('format', None):
+                hotspot_columns[entry]['format'] = columns['columns'][entry]['format']
         if 'order' in columns['columns'][entry]:
-            if entry in hotspot_columns:
-                hotspot_columns.remove(entry)
             output_order.append((columns['columns'][entry]['order'], entry))
             if 'from' in columns['columns'][entry] and 'merge' == columns['columns'][entry]['from']:
                 report_header.append((columns['columns'][entry]['order'],
@@ -93,9 +105,53 @@ def generate_hotspot_report(sample,
             else:
                 report_header.append((columns['columns'][entry]['order'], entry))
             biggest_order = max(biggest_order, columns['columns'][entry]['order'])
-        elif columns['columns'][entry].get('visible', 1) == 0:
-            if entry in hotspot_columns:
-                hotspot_columns.remove(entry)
+    for key in hotspot_columns:
+        if key in columns['columns']:
+            del columns['columns'][key]
+
+    def handle_select(data):
+        def convert_list_slice(info):
+            if len(info) < 3:
+                raise SyntaxError(f"Invalid syntax: {info}")
+            if info.startswith("[") and info.endswith("]"):
+                info = info[1:-1].split(":")
+                if len(info) == 1:
+                    return lambda data: data[int(info[0])]
+                elif len(info) == 2:
+                    value0 = int(info[0]) if len(info[0]) > 0 else None
+                    value1 = int(info[1]) if len(info[1]) > 0 else None
+                    return lambda data: data[value0:value1]
+                elif len(info) == 3:
+                    value0 = int(info[0]) if len(info[0]) > 0 else None
+                    value1 = int(info[1]) if len(info[1]) > 0 else None
+                    value2 = int(info[2]) if len(info[2]) > 0 else None
+                    return lambda data: data[value0:value1:value2]
+                else:
+                    raise SyntaxError(f"Invalid syntax: {info}, invalid number of elemens inside '[]'")
+            else:
+                raise SyntaxError(f"Invalid syntax: {info}, missing brackat '[]'")
+
+        def condition(items, empty):
+            items = convert_list_slice(items)
+
+            def func(data):
+                values = items(data)
+                if isinstance(values, str):
+                    return (values == "-") == empty
+                else:
+                    return (len([v for _, v in values.items() if v != "-"]) == 0) == empty
+            return func
+
+        data['items'] = convert_list_slice(data['items'])
+        if "else" in data:
+            data['else'] = convert_list_slice(data['else'])
+        if "condition" in data:
+            data['condition'] = condition(data['condition']['items'], data['condition'].get("empty", True))
+        return data
+
+    for entry in columns['columns']:
+        if "from" in columns['columns'][entry] and "select" in columns['columns'][entry]["from"]:
+            columns['columns'][entry] = handle_select(columns['columns'][entry])
 
     for entry in hotspot_columns:
         biggest_order += 1
@@ -156,6 +212,7 @@ def generate_hotspot_report(sample,
                                     'gvcf_depth': depth,
                                     'ref_depth': '-',
                                     'alt_depth': '-'}
+                            format_hotspot(data, columns, hotspot_columns)
                             add_columns(data, None, hotspot, columns, annotation_extractor, depth, levels)
                             writer.write("\n" + "\t".join([str(data[c[1]]) for c in output_order]))
                             counter += 1
@@ -173,6 +230,7 @@ def generate_hotspot_report(sample,
                                     'gvcf_depth': depth,
                                     'ref_depth': var.samples[sample]['AD'][0],
                                     'alt_depth': ",".join(map(str, var.samples[sample]['AD'][1:]))}
+                            format_hotspot(data, columns, hotspot_columns)
                             add_columns(data, var, hotspot, columns, annotation_extractor, depth, levels)
                             writer.write("\n" + "\t".join([str(data[c[1]]) for c in output_order]))
                             counter += 1
@@ -192,10 +250,37 @@ def generate_hotspot_report(sample,
                     'gvcf_depth': depth,
                     'ref_depth': var.samples[sample]['AD'][0],
                     'alt_depth': ",".join(map(str, var.samples[sample]['AD'][1:]))}
+            format_hotspot(data, columns, hotspot_columns)
             add_columns(data, var, None, columns, annotation_extractor, depth, levels)
             writer.write("\n" + "\t".join([str(data[c[1]]) for c in output_order]))
             counter += 1
             log.info("-- non-hotspot entries: {}".format(counter))
+
+
+def format_value(value, format):
+    if format[0] == "replace":
+        return value.replace(format[1], format[2])
+    elif format[0] == "string":
+        if len(format) == 3:
+            value = getattr(builtins, format[2])(value)
+        else:
+            value = float(value)
+        return format[1].format(value)
+    else:
+        raise Exception("Unknown format value: " + format)
+
+
+def format_hotspot(data, columns, hotspot):
+    for key in data:
+        format = columns.get(key, {}).get("format", None)
+        format = hotspot.get(key, {}).get("format", format)
+        if format:
+            try:
+                data[key] = format_value(data[key], format)
+            except ValueError:
+                log.warning("Unable to format value {}, field {}, format {}".format(data[key], key, format))
+            except TypeError:
+                log.warning("Unable to format value {}, field {}, format {}".format(data[key], key, format))
 
 
 def extract_item_merge_header(columns):
@@ -215,11 +300,36 @@ def add_columns(data, var, hotspot, columns, annotation_extractor, depth, levels
     """
     def add_data(data, column, c, depth, levels):
         if column[c]['from'] == "merge":
+            # Used to merge two columns and combine using divider
             divider = column[c]['divider']
-            temp_data = {}
+            temp_data = OrderedDict()
             for key in column[c]['elements']:
                 add_data(temp_data, column[c]['elements'], key, depth, levels)
             data[c] = divider.join([v for k, v in temp_data.items()])
+        elif column[c]['from'] == "select":
+            # From multiple column select the first one with data
+            temp_data = OrderedDict()
+            for key in column[c]['elements']:
+                add_data(temp_data, column[c]['elements'], key, depth, levels)
+            temp_data = [v for _, v in temp_data.items()]
+            select = column[c]['items'](temp_data)
+            if not isinstance(select, str):
+                select = column[c].get("divider", ":").join(select)
+            data[c] = "-"
+            condition = True
+            if 'condition' in column[c]:
+                condition = column[c]['condition'](temp_data)
+            if select is not None and select != '-' and condition:
+                data[c] = select
+            elif 'else' in column[c]:
+                select = column[c]['else'](temp_data)
+                if not isinstance(select, str):
+                    select = [v for v in select if v != "-"]
+                    if len(select) == 0:
+                        select = "-"
+                    else:
+                        select = column[c].get("divider", ":").join(select)
+                data[c] = select
         elif column[c]['from'] == "vep":
             try:
                 data[c] = annotation_extractor(var,  column[c]['field'])
@@ -241,24 +351,32 @@ def add_columns(data, var, hotspot, columns, annotation_extractor, depth, levels
                 variables = []
                 for v in column[c]['variables']:
                     variables.append(locals().get(v, v))
-                value = "-"
                 try:
                     value = function(*variables)
                 except AttributeError:
-                    data[c] = "-"
+                    value = "-"
             else:
-                function()
+                value = function()
             if 'column' in column[c]:
                 data[c] = value[column[c]['column']]
             else:
                 if isinstance(value, tuple):
                     value = ",".join(value)
                 data[c] = value
+            if data[c] is None:
+                data[c] = "-"
         elif column[c]['from'] == 'variable':
             data[c] = locals()[column[c]['field']]
         else:
             raise Exception("Undhandledd cased: " + column[c]['field'])
+        if "format" in column[c]:
+            try:
+                data[c] = format_value(data[c], column[c]["format"])
+            except ValueError:
+                log.debug("Unable to format value {}, field {}, format {}".format(data[c], c, column[c]["format"]))
+            except TypeError:
+                log.debug("Unable to format value {}, field {}, format {}".format(data[c], c, column[c]["format"]))
 
-    for c in columns["columns"]:
-        if 'from' in columns["columns"][c]:
-            add_data(data, columns["columns"], c, depth, levels)
+    for c in columns['columns']:
+        if 'from' in columns['columns'][c]:
+            add_data(data, columns['columns'], c, depth, levels)